Lv2

+++ src/main/java/com/sparta/lv2blogspring/service/PostService.java
@@ -1,63 +1,74 @@
 package com.sparta.lv2blogspring.service;
 
+import com.sparta.lv2blogspring.dto.PostListResponseDto;
 import com.sparta.lv2blogspring.dto.PostRequestDto;
 import com.sparta.lv2blogspring.dto.PostResponseDto;
 import com.sparta.lv2blogspring.entity.Post;
+import com.sparta.lv2blogspring.entity.User;
 import com.sparta.lv2blogspring.repository.PostRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.concurrent.RejectedExecutionException;
+import java.util.stream.Collectors;
 
 @Service
+@RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
 
-    public PostService(PostRepository postRepository) {
-        this.postRepository = postRepository;
-    }
+    public PostResponseDto createPost(PostRequestDto requestDto, User user) {
+        Post post = new Post(requestDto);
+        post.setUser(user);
 
-    public List<PostResponseDto> getPostList() {
-        return postRepository.findAllByOrderByCreateAtDesc().stream() //
-                .map(PostResponseDto::new).toList();
-    }
-    //DB조회 리스트 -> 스트림 -> PostResponsDto로 변환 -> Stream 다시 List 변환
+        postRepository.save(post);
 
-    public PostResponseDto getPost(Long id) {
-        Post post = findPost(id);
         return new PostResponseDto(post);
-
     }
 
-    public PostResponseDto createPost(PostRequestDto postRequestDto) {
-        Post post = new Post(postRequestDto);
-        Post savePost = postRepository.save(post);
-        return new PostResponseDto(savePost);
+    public PostListResponseDto getPosts() {
+        List<PostResponseDto> postList = postRepository.findAll().stream()
+                .map(PostResponseDto::new)
+                .collect(Collectors.toList());
+
+        return new PostListResponseDto(postList);
+    }
+
+    public PostResponseDto getPostById(Long id) {
+        Post post = findPost(id);
+
+        return new PostResponseDto(post);
+    }
+
+    public void deletePost(Long id, User user) {
+        Post post = findPost(id);
+
+        if (!post.getUser().equals(user)) {
+            throw new RejectedExecutionException();
+        }
+
+        postRepository.delete(post);
     }
 
     @Transactional
-    public PostResponseDto updatePost(Long id, PostRequestDto postRequestDto) {
+    public PostResponseDto updatePost(Long id, PostRequestDto requestDto, User user) {
         Post post = findPost(id);
-        post.checkPassword(postRequestDto.getPassword());
 
-        post.setTitle(postRequestDto.getTitle());
-        post.setName(postRequestDto.getName());
-        post.setContent(postRequestDto.getContent());
+        if (!post.getUser().equals(user)) {
+            throw new RejectedExecutionException();
+        }
+
+        post.setTitle(requestDto.getTitle());
+        post.setContent(requestDto.getContent());
 
         return new PostResponseDto(post);
     }
 
-    public void deletePost(Long id, String password) {
-       Post post = findPost(id);
-       post.checkPassword(password);
-
-       postRepository.delete(post);
-    }
-    private Post findPost(Long id) {
+    private Post findPost(long id) {
         return postRepository.findById(id).orElseThrow(() ->
-                new IllegalArgumentException("해당 글이 존재하지 않습니다."));
+                new IllegalArgumentException("선택한 게시글은 존재하지 않습니다.")
+        );
     }
-
 }
-
-
